Whitespace cleanup and server detection changes
Converted tabs to spaces in various spots and adjusted to 4-space indentation (a project that will continue as time goes on).

Also removed old discovery code which has been replaced with Cayce's multicast discovery code. Also, when adding a server by name, store the ip address in the URL field instead of storing the hostname. This resolves issues with transcoding. Affected users will need to remove and re-add the server (either manually or via discovery).