A grid view inspired by L0st-S0ul's PlexGrid.
A working, Netflix-style grid view, inspired by L0st-S0ul's awesome work on
PlexGrid. This is very much based on that project, but not a straight merge.
The main difference is that instead of loading an entire row at a time, we
make paginated requests within rows. This ought to help for extremely large
sections. Besides that, it's mostly just structured a little differently,
wrapping things up as an object.