Revamped poster screen.
Wrap the poster screen up as an object, just like the rest of our new screens.
The way poster content is loaded is also new. We used to try to load large
sections in chunks by only ever having 11 items loaded at a time. This meant
that every time you selected a new item, we'd have to refresh the list from
the server to include the window of 11 items. Instead, we now load in chunks,
but we remember everything we've already loaded and simply append whatever
we just loaded. This makes navigation feel snappier without sacrificing
anything in the initial load time.