Further work on GHI #33
Add the ability to move back and forth in photos and rate them. The photo section is shown in a grid view but folders are shown in poster view (as the grid parsed each photo as it's own line, which looked silly). Photo elements that do not end in "children" are sent to the springboard when selected (as they are not folders). (I could not find a good way to distinguish between directory elements and photo elements the way we handle XML parsing.) This could use a second pair of eyes, but it works fine and does support nested folders correctly.