Load home screen content asynchronously.
Instead of blocking while loading a particular request, kick off all the
requests for a given row at the same time and handle the responses in the
screen's event loop. This should make the home screen more responsive in
general, but especially for remote servers, which will be more important
once we're showing myPlex content on the home screen.

This is most of the battle for #42, but we're still validating servers
before doing anything else, so this doesn't fix that.