Add token to image and music requests.
For servers that have an access token, add the token on all requests for
metadata (usually posters and MP3s). Usually we can add it as a header since
all screens implement ifHttpAgent, but on the home screen we have to add it
as a query string parameter since we're showing content from multiple servers.
Part of #20.