Universal search. Fixes #10.
Add a single search item to the Misc row, and use it to kick off a universal
search across all owned servers (myPlex or locally configured). The results
are shown in a grid with a row for each of the types that search returns
(e.g. Movies, Episodes, Shows). The search results are populated
asynchronously.