Enable slightly lame search suggestions on universal search.
See #46.

This is a little lame at the moment because it's not obvious who we should
ask for search suggestions. When searching within a section, we're only ever
going to search a single server/URL, so we just hit it with a paginated
request. For universal search, we may search several servers, and even
within each server we may follow several URLs (e.g. searching channel
content).

At the moment, we just arbitrarily choose a single owned server, and pull
suggestions from its top level search (/search?query=blah). I'm not sure
if this is better than nothing, or if showing empty suggestions on certain
searches will make some users think that there will be no results if they
hit search.