Try to improve performance of poster screens.
The poster screens don't perform as well as grid screens, and one of the
reasons is that there's no method for updating only a portion of the screen's
content list. So, instead of updating the list every time a page of content
comes in, update our backing data but only tell the screen about the new
data when the user has focused an item in the vicinity of the end of the list.

This also fixes a couple things from the recent data loader changes.