Initial code for queue on home screen.
Part of #20. The calls to actually initialize the queue row are still commented
out, but if you uncomment them you'll see the queue items show up correctly on
the home screen. It's still commented out because you can't play most (all?)
items.

Also, this implements a temporary notion of a primary server, which is simply
an arbitrary owned server. We need some way to decide which server to use for
transcoding queue content that can't be direct played (which is everything
until we support direct play).