Move around some home screen initialization code.
In addition to deleting a bunch of duplicated stuff, the goal here is to
make it easier to refresh the home screen. Depending on what you change in
the preferences, the home screen may or may not refresh correctly, but even
when it refreshes it does so expensively. If I go into the preferences and
sign in to myPlex, there's no need to refresh all of the configured servers.
This change doesn't improve that yet (it actually breaks all refreshing), but
it should make it easy to support that.