Fix for black grid screen after playing audio.
As reported by L0st-S0ul (and fixed in his PlexAudio), after using an audio
player and backtracking, the grid becomes corrupted and just shows as black.
The fix is to close the grid screen and on the way back create a new one
with the same content. This isn't actually too awkward since we're already
wrapping the actual grid screen.