WTF? Don't crash when Roku sends us bogus event.
If I do a universal search and immediately select the first thing focused,
our grid screen is getting a list item focused event for row 20, which doesn't
exist. I don't see how we could be causing this event, but it's possible. At
any rate, we need to ignore it to avoid crashing.