Allow individual screens to specify a theme.
Some of the Roku screens aren't really customizable enough to support a
dark background color. The search screen, for example, won't let you
change all of the colors, so using a background that matches the grid
screen's default looks kind of terrible.

It might be smart to just always use the defaults, but this allows us to
change the backgrounds and colors on a per-screen basis. The only way we're
really using it at the moment is to show a dark background for poster
screens. So, generally speaking all of your browsing will be against a
dark background, and then when you get to the springboard, or a search
screen, you get the default scheme, which tends to look better.

The springboard is at least pretty configurable, but in a pretty daunting
way, so I haven't tried to figure out what colors would look good for
all the different elements against a dark background.