Allow direct play to be forced on a per-item basis.
Create a playback options screen for video items. Move the audio and subtitle
selection to that screen, and introduce an option to choose between Auto,
Force Direct Play, Direct Play with Fallback, and Transcode. (The difference
between Auto and DP w/ Fallback is that Auto will only try DP if it thinks
it will work.)