Try to fix a terrible audio player crash.
If we give the audio player a bunch of items with no URL, it will sometimes
panic and render the device useless (no apps work until a hard reset). As long
as we make the URL "" and the codec something bogus, it seems to be ok if the
entire list isn't bogus. So, the attempt here is basically two-fold:

1) Make sure the URL is never invalid.
2) If nothing in the list is playable, don't even try to play. Don't show a
   springboard, just show an error.

Seems to do the trick for me, hopefully fixes #71.