Refresh grid rows after coming back from another screen. Fixes #58
This works ok. The problem is we don't really want to start from scratch on
the way back since things are probably mostly the same. But if we don't
start from scratch there's not really any indication that things are loading
in the background. It works pretty well if you stay in the top rows though.
For example, refreshing an On Deck row in a TV section where the On Deck
list isn't huge is pretty snappy.

We're not refreshing anything but the queue on the home screen. This is not
necessarily optimal, but the alternative is a bit tricky. We also don't
refresh anything on search results grids.