Don't Direct Stream to Roku1 by default.
The Roku1 is picky about H.264 streams in HLS. I have some files that I can
package as MP4 and they'll Direct Play fine, but if I Direct Stream that
exact same file then the playback is terrible. (And it Direct Streams fine
on my Roku2.) We can't really predict when this will happen, so by default
we're not going to Direct Stream to a Roku1 anymore.