Don't crash when moving left off the end of the queue.
We allow left/right navigation while on springboard screens, but in the queue
we have a dummy item that isn't a video (to pull up the rest of the queue).
Trying to refresh that item on a springboard was crashing. Filter out any
items that can't be shown on a springboard when setting the springboard's
context.