Bump build number.
Kind of annoyingly to those that have followed along, I'm calling this a 2.0
build, not 2.1. Since the public channel was a brand new channel in Roku's
eyes, it had to be called 1.0 in the store even though it's 2.x to us. Once we
make an update, we can call it 2.0, but we can't call it 2.1 until we have a
2.0 package approved. And in the eyes of the users we're hoping to release an
update within the week of the original release, so being an iteration of 2.0
instead of a jump to 2.1 isn't that crazy.