Shuffle support for audio playback, fixes #91.
Add support for shuffling (and unshuffling) the current context being played
by the audio player.

In order to do this, we need more than 5 buttons on the springboard. So we now
have a "more..." option that pulls up rating and shuffle. There are a couple
more things on the horizon that make sense on this dialog, so it's not a total
waste.