Use longer HLS segments, mitigate Roku 2 HLS woes.
A variety of Roku 2 quirks with HLS appear to relate to stitching together
segments. Maybe it's a Roku bug, maybe we should be generating .ts segments
that stand on their own (e.g. always start with I-Frames). It's not really
clear yet.

In the meantime, using 10s segments as opposed to 1s segments gets us the
following:

* Noticeably improved load times, even on 4.6. With the 4.6 firmware, I've
  observed streams loading about four times faster.
* Dramatically reduced picture issues, #95. If the artifacts only occur at
  segment boundaries (which seems to be true), then it should be 10x better
  now. Anecdotally, it seemed even better than that on the videos I tried.
* Possibly reduced audio glitches with 5.1 sound, #96. I don't have a setup
  that reproduces the problem, but a separate test suggested that the audio
  problems were less bad when lowering the transcode quality, even though
  the audio stream was copied in both cases. Since the audio stream was
  copied, the only relevant difference should have been the longer segment
  size with the lower quality.