Improve selection of preferred media item, fixes #116.
Use the exact same "can direct play" logic when selecting a preferred media
item so that we never fail to pick an item that could have direct played. If
there are no items that can direct play, choose based on a trivial scoring
system that prefers streams that might support copying. The "can direct play"
decision is also cached on the item so that we don't run through the logic
more than necessary.