Allow row order to be customized with a preference.
Read a (presently hidden) preference instead of hardcoding section keys that
will be moved to the top of the grid. This provides an outlet other than
forking for users who really want something else to be at the top of the list,
without messing with how a user might actually configure such a thing in the
UI (which I'm still not sure we ever want to do).

To set it using the launch params, pass a comma-separated list of keys. The
comma has to be encoded though, so it's really a `%2c` separated list. The
last key will show up as the first row. For example:

```
curl -XPOST 'http://10.0.1.2:8060/launch/13535?pref!priority_keys=newest%2CrecentlyAdded%2CrecentlyViewed%2ConDeck%2Call'
```

or

```
curl -XPOST 'http://10.0.1.2:8060/launch/13535?pref!priority_keys=newest%2CrecentlyAdded%2CrecentlyViewed%2ConDeck%2Call'
```

Note that the IP address needs to be the IP of your Roku, and the number after
launch in the URL is the channel ID. You can use 13535 for the official
channel, 14987 for PlexTest, or dev for a dev channel.