Fix stupid crash in playback options screen.
Don't define a function as returning a string and then return invalid. I'm a
little surprised we've never run into this, but I specifically hit it when
trying to pull up the playback options for an RTMP channel. I didn't really
investigate if there were mitigating factors to explain the lack of reports,
the code was clearly wrong.