Initial work toward a global message port/loop.
Create a single message port to be shared in the view controller, and
run a main message loop in the controller. Eventually the idea is to
just have this single message loop, but it's a massive change and this
is just the beginning. The basic grid screen and simple data loaders
have been updated to use the new methods. Everything else, including
the home screen and the grid screen with search data loader, is still
pending.