Get the home screen working with the global message loop.
This is a fairly big and horrifying change, but it's actually made up
of several parts that hopefully make sense.

First, instead of the home screen being a confusing combination of a
Screen and DataLoader, whose Screen is actually another wrapped screen
rather than a native screen, the responsibilities have been separated.
The home screen is now a regular GridScreen with just a couple tweaks,
like changing the style and up behavior. All of the trickiness was always
in the custom data loader, and so that has been separated into its own
object/file.

Second, in order to handle GDM in the global message loop we need a way
to listen for socket events. This works basically the same way as URL
events. Objects register to be notified of socket events on a particular
socket; the view controller looks for a registered listener when it
receives a socket event.

Third, in order to show the help screen after several seconds of hoping
for GDM to find a server, we need a timer concept. This also works
somewhat similarly to the URL events. The old performance timer abstraction
has been generalized into a Timer object. Objects can register timers
with the view controller. On every event, the view controller checks
to see if any timers have expired and notifies the object that registered
the timer. The view controller is also able to check all of the timers
in order to set an appropriate timeout on the global wait call.