Convert the photo springboard and player to use the global message port.
There has always been a lot of shared code for the three springboards, but it's
always been a bit messy (my fault). This tries to clean that up a bit. There's
now a much more explicit base springboard screen which will be extended by the
photo/audio/video varieties. The players will also be distinct from the
springboards in this enlightened new world. We're not taking advantage of that
yet, except to the extent that it's more consistent to have one of our "Screen"
wrappers around an roSlideshowScreen, but to use a simple example there's no reason
not to launch a slideshow directly from the grid screen if the user hits the
Play button on the remote. Previously that was all but impossible, but it should
be quite trivial after separating the player from the springboard.