Rework audio springboard and player to use global message port.
This is another big change, but it's consistent with the idea of the
rest of the global message port changes. Continuing on the work for the
photo screens, the audio springboard extends the base springboard
and is a fairly normal screen. The audio player is a separate object
and mostly independent of the springboard screen. Among other things,
this will allow background audio playback. Sort of as proof of that,
the channel will currently leave the music playing when you leave the
springboard. This is far from ideal as-is though, as there's no way
to bring up audio player controls once you leave the springboard
screen.