Improved perceived load time on grid screens.
When we're loading grid rows, we generally know how big the row will be,
but since we're loading in chunks the size keeps growing. It also means that
moving left off the beginning of the data will take you to the current last
element, which quickly becomes a random middle element. Instead, we now
immediately make the array its eventual full size so that the grids show
the proper counts and don't bounce around. Pending items show themselves
as loading.