Provide a way to get back to the currently playing audio.
This is part of #45. Specifically:

- Only process audio player events on the audio springboard that's responsible
for the audio player. This allows you to start playing music and continuing
browsing music without it automatically stopping the current music and playing
whatever you're browsing.

- When the audio player has a current "playlist", show a Now Playing item on
the home screen. Selecting this item brings up a new springboard screen with
whatever selection of music the audio player is going through.