Show an audio player context menu on * presses, #45.
Also part of #68.

When the user presses *, if we have a current playlist in the audio player,
show a dialog with audio controls, including an option to go to the now playing
screen.

As delightful as this is, note that we're at Roku's mercy here. The * key can
only be trapped on certain screens. For the screens we use though, it'll work
on all of the poster and grid screens, as well as search and list screens, so
it's mostly just an issue on springboard screens.