Handle "Stop" remote command specially, #52.
Instead of assuming the user is hitting Stop in a scenario where treating it
like "Back" makes sense, check to see if a video is playing. If a video is playing,
close the current screen. Otherwise tell the audio player to stop. If neither a
video nor music is playing then it's a no-op.