A lame but mostly functional multi-part video implementation, #101.
If we played a video (er, part, to be precise) and there are more parts for
the current media item, throw up another video player for the next part. This
works fairly well in the simple cases, which includes some of the popular
channels that have gone to stacked MP4s instead of WebKit transcoding. It
doesn't really work at all in the face of stream selection.