Add pref to hide the channels section
I just followed the model for the other sections that can be hidden.
One quirk that I noticed was that whenever any section is hidden, an
invisible row appears at the top of the home screen. Exiting Plex on
the Roku and reentering fixes the issue. This quirk exists for all of
the rows that can be hidden, not just Channels.