Ping the transcoder every minute, no matter what.
We can't guarantee that the PMS is receiving requests that will keep the
transcode session alive during normal playback, so ping the transcoder every
minute. I don't have specific evidence that this has been a problem, but a
couple of reports have me suspicious, and this doesn't really hurt anything
anyway.