Work around Roku 3.x audio player and grid screen glitch, fixes #183.
This is actually much uglier than it used to be. We used to have it easy
since audio could never be playing on a grid screen. Now that it can, it
turns out that the glitch isn't caused by creating an audio player, as
previously thought, but any time the audio player starts a new track. So
we need to destroy grid screens every time a new track starts, and then
recreate them when that screen would be activated.

If the current screen is a grid screen, then we need to additionally
make sure we ignore the screen closed event, and we can't call activate
to recreate the grid right away, so we need to do it on a timer. Really
ugly.