Improve direct play -> transcode fallback.
If we were already transcoding, whether it was a remux or transcode, don't
fall back to anything. This was causing users to report that part way through
videos would restart from the beginning. What was actually happening is that
the transcoder was dying, but then we tried the same thing again and it would
fail in the same place.

And if we fall back after having successfully played some amount of the video,
restart from that offset rather than the original seek value.