Wait for a segment to be available before really starting, fixes #174.
The Roku does its buffering math for HLS based on how quickly it's able to
pull down the first segment. This works against us, since we're creating that
segment on the fly and there's some lag before PMS returns it. We appear to
be able to make the Roku use a more realistic number if we ask the server
not to return the M3U8 (and therefore really start the video player) until
the first segment is really available.