Add a separate quality pref for non-local content, fixes #207.
This also requires a (welcome) reworking of the way quality is temporarily
overridden. We no longer write to the main quality pref to do an override
and then try to restore it later. Instead we have a separate registry key
for the override, and delete that override when we're done with it (including
at startup, to make sure it can't linger from a previous run).