Add lame but effective continuous play mode, #115.
We're still not using `roVideoPlayer`, so we can't use the content list feature
of that video player. We're also still not implementing playlists, so there's no
way to select that an arbitrary set of things be played continuously. But it's
now possible to enable continuous play and when the video is played from a springboard
screen it'll automatically start playing the next video in the context list.