For myPlex content, fetch details from node, not myPlex.
myPlex appears to send back full media info, but it might not be accurate. But
requesting the info from node means that we no longer see updated info about things
like viewOffset when refreshing. So refresh using the myPlex handcrafted detail
URL, and fetch media using the plain old key which happens to point to node.