Switch to /:/timeline for reporting progress, fixes #50.
It wasn't even that hard, welcome to the modern world!

We basically just use a couple of timers. We still get notified by the Roku
video player during playback and we send those events as "playing" timeline
events. We also keep track of our play state and send an event whenever it
changes. Finally, we use a timer to make sure we continue to send events
(every 15 seconds) if there hasn't been a natural event. This ensures we
keep sending timeline events while paused (or buffering).

The only thing we don't really try to track is buffering. There's no event to
tell us that buffering is happening. If we wanted to track it, we'd probably
need to crank up both our own timer and the Roku progress timer, so that we'd
know for example "it's been 2 seconds and we didn't get our every second progress
notification, must be buffering".

myPlex doesn't provide a /:/timeline endpoint yet, so we translate those requests
to the old progress and scrobble endpoints.