Tweak totally bogus exif check.
This check is totally bogus, but we don't have exif information in the XML yet
and this helps when browsing from iOS photos. The width vs. height check was the
reverse of what I meant to check, but it's not even going to be right that often.
And if iOS adds width/height before adding some exif information to key off of,
the photos are going to be transcoded for being unnecessarily large anyway. So
it seems fine all things considered to transcode anything for which we have no
information.