Check device info for surround support periodically, fixes #229.
The Roku 3 surround support can be a moving target because of the headphones in
the remote. So we can't just check at startup and remember the value anymore.
But reading the device info was observed to be slow, so we don't want to check
every time in the loop where we process video item XML. We use a simple timer
to refresh the value periodically, and make sure we have the current value before
we start to actually play a video.