Try to differentiate between channel search and text input.
The Framework doesn't do us any favors here. Channel developers say they want
some text input and the XML says it's a search item. If they say they want a
search, that's also a search item in the XML. On the Roku, we now clumsily
try to differentiate based on the key (is it a search service?) and the
prompt. If it looks like a text input, then we use a keyboard screen instead
of the search screen. The keyboard allows for more characters and has no notion
of progressive search, so incomplete text won't be sent to the channel.

In order to pull it off it was also necessary to tweak how posters decide if
they should show a message about empty containers.

And in other news, we're down to one available function again. :[