Add clock as breadcrumb text on home screen. h/t @ljunkie
This has been requested a handful of times and recently showed up in
the rarflix fork. We don't have anything else to show up there and it
seems to have received a positive response.

Since we're only showing it on the home screen, we run the timer all
the time but only actually resolve the current time when the home screen
is visible.