This module holds the multiple-user support for Plex
Multi-user support is implemented by creating separate preference sections for
each user profile.  There are Reg*() routines that handle all this automatically
for you.  Just call the usual RegRead() and RegWrite() and they
will read/write to the correct preference.

The "default user" or "User 0" is special, in that the preference sections
for that user are the same as they are without multiple users.  This makes
upgrading easy.  For Users1-3, their preference sections are the usual section
names but with a "_uN" added onto them.  But again, the Reg*() routines will
handle this conversion.

Multi-user also allows a security PIN (a sequence of the directional arrows)
for any of the user profiles.

Limitations:
*   Switching users requires restarting the channel
*   User number is limited to 4.  That is a limit set by the desire to have a single
    button user selection (the direction arrows) on start-up.  That can easily be increased
    without limit with a different user selection screen.