Implemented working Wake-On-LAN Functionality
This is the final part of Wake-On-LAN (WOL) functionality for Roku Plex.
It adds an extra setting to the Server settings - an optional "SecureOn"
password that is used for WOL.

Before making an HTTP request for PMS's library, Roku now checks to see
if there are any WOL settings saved.  If there are, it first sends 5
Magic packets to wake up the server.  These packets are sent over 500ms
and are synchronous (therefore hold up the boot process).

From my testing, there has always been enough time for the HTTP request
to successfully reach the server after a wake-up.  This might need to be
optimized for slower servers.